+++ packages/react-kit/src/components/EventListener/EventListener.tsx
@@ -51,7 +51,9 @@ export class EventListener extends React.Component<TEventListenerProps> {
 			const capture = key.endsWith(CAPTURE_MARKER);
 			const handler = handlers[key];
 			const eventName = getEventName(key, capture);
-			target.addEventListener(eventName, handler as any, capture);
+			const isTouchEvent = ['touchcancel', 'touchend', 'touchmove', 'touchstart'].includes(eventName);
+			const options = isTouchEvent ? { capture, passive: false } : { capture };
+			target.addEventListener(eventName, handler as any, options);
 		});
 	}