+++ packages/react-kit/src/components/RootClose/RootClose.tsx
@@ -61,9 +61,10 @@ export class RootClose extends Component<TRootCloseProps> {
 		this.preventMouseRootClose = domNode.contains(e.target as Node);
 	};
 
-	private handleTouchStart: TouchEventHandler<HTMLElement> = () => {
+	private handleTouchStart: TouchEventHandler<HTMLElement> = e => {
 		if (!this.props.ignoreClick && !this.preventMouseRootClose && this.props.onRootClose) {
 			this.props.onRootClose();
+			e.preventDefault();
 		}
 	};