+++ packages/rx-utils/src/live-data.utils.ts
@@ -1,4 +1,4 @@
-import { Observable, of, combineLatest } from 'rxjs';
+import { Observable } from 'rxjs';
 import { remoteData, RemoteData, failure, isSuccess } from '@devexperts/remote-data-ts';
 import { Option } from 'fp-ts/lib/Option';
 import { MonadThrow2 } from 'fp-ts/lib/MonadThrow';
@@ -6,15 +6,29 @@ import {
 	CoproductLeft,
 	coproductMapLeft,
 } from '@devexperts/utils/dist/typeclasses/product-left-coproduct-left/product-left-coproduct-left.utils';
-import { getApplicativeComposition } from 'fp-ts/lib/Applicative';
 import { instanceObservable } from './observable.utils';
 import { pipeable } from 'fp-ts/lib/pipeable';
 import { sequenceT } from 'fp-ts/lib/Apply';
 import { array } from 'fp-ts/lib/Array';
+import { Filterable2 } from 'fp-ts/lib/Filterable';
+import { getLiveDataM } from '@devexperts/utils/dist/adt/live-data.utils';
+import { FoldableValue2 } from '@devexperts/utils/dist/typeclasses/foldable-value/foldable-value';
 
+/**
+ * isomoprhic to `LiveData12<typeof instanceObservable.URI, typeof remoteData.URI, E, A>`
+ * but left as an interface to not overload typechecker and IDE
+ *
+ * @deprecated Compose `LiveData` in the end project
+ */
 export interface LiveData<E, A> extends Observable<RemoteData<E, A>> {}
 
-export const URI = '@devexperts/dx-utils//LiveData';
+/**
+ * @deprecated Compose `LiveData` in the end project
+ */
+export const URI = '@devexperts/rx-utils//LiveData';
+/**
+ * @deprecated Compose `LiveData` in the end project
+ */
 export type URI = typeof URI;
 declare module 'fp-ts/lib/HKT' {
 	interface URItoKind2<E, A> {
@@ -22,19 +36,23 @@ declare module 'fp-ts/lib/HKT' {
 	}
 }
 
-const tupleC = <A>(a: A) => <B>(b: B): [A, B] => [a, b];
-const combineRD = <EA, A, EB, B>(fa: RemoteData<EA, A>, fb: RemoteData<EB, B>): RemoteData<EA | EB, [A, B]> =>
-	remoteData.ap<EA | EB, B, [A, B]>(remoteData.map(fa, tupleC), fb);
-
-export const instanceLiveData: MonadThrow2<URI> & CoproductLeft<URI> = {
-	URI,
-	...getApplicativeComposition(instanceObservable, remoteData),
-	chain: (fa, f) => instanceObservable.chain(fa, a => (isSuccess(a) ? f(a.value) : of(a))),
-	throwError: e => of(failure(e)),
-	coproductLeft: <EA, A, EB, B>(fa: LiveData<EA, A>, fb: LiveData<EB, B>) =>
-		instanceObservable.map(combineLatest(fa, fb), ([a, b]) => combineRD(a, b)),
+const foldableValueRemoteData: FoldableValue2<typeof remoteData.URI> & MonadThrow2<typeof remoteData.URI> = {
+	...remoteData,
+	foldValue: (fa, onNever, onValue) => (isSuccess(fa) ? onValue(fa.value) : onNever(fa)),
+	throwError: failure,
 };
 
+/**
+ * @deprecated Compose `instanceLiveData` in the end project
+ */
+const instanceLiveData: MonadThrow2<URI> & CoproductLeft<URI> & Filterable2<URI> = {
+	URI,
+	...getLiveDataM(instanceObservable, foldableValueRemoteData),
+};
+
+/**
+ * @deprecated Compose `liveData` in the end project
+ */
 export const liveData = {
 	...instanceLiveData,
 	...pipeable(instanceLiveData),
@@ -43,4 +61,7 @@ export const liveData = {
 	combine: coproductMapLeft(instanceLiveData),
 };
 
+/**
+ * @deprecated Compose `LiveDataOption` in the end project
+ */
 export interface LiveDataOption<E, A> extends Observable<RemoteData<E, Option<A>>> {}