+++ packages/rx-utils/src/observable.utils.ts
@@ -2,12 +2,13 @@ import { observable as rxjs, URI } from 'fp-ts-rxjs/lib/Observable';
 import { MonadObservable1 } from '@devexperts/utils/dist/typeclasses/monad-observable/monad-observable';
 import { Observable, Subject } from 'rxjs';
 import { pipe, pipeable } from 'fp-ts/lib/pipeable';
-import { share } from 'rxjs/operators';
+import { share, shareReplay, switchMap } from 'rxjs/operators';
 import { fromPromise } from 'rxjs/internal-compatibility';
 import { task } from 'fp-ts/lib/Task';
 
 export const instanceObservable: typeof rxjs & MonadObservable1<URI> = {
 	...rxjs,
+	chain: (fa, f) => pipe(fa, switchMap(f)),
 	createAdapter: <A>() => {
 		const s = new Subject<A>();
 		const next = (a: A) => s.next(a);
@@ -47,8 +48,10 @@ export const instanceObservable: typeof rxjs & MonadObservable1<URI> = {
 };
 
 const zero: () => Observable<never> = instanceObservable.zero;
+const hold: <A>(source: Observable<A>) => Observable<A> = shareReplay({ refCount: true, bufferSize: 1 });
 export const observable = {
 	...instanceObservable,
 	...pipeable(instanceObservable),
 	zero,
+	hold,
 };